requirements.txt
beautifulsoup4==4.9.1
boto3==1.14.49
botocore==1.17.49
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
docutils==0.15.2
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.10
jieba3k==0.35.1
jmespath==0.10.0
joblib==0.16.0
lxml==4.5.2
newspaper3k==0.2.8
nltk==3.5
Pillow==7.2.0
pkg-resources==0.0.0
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-file==1.5.1
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==2.2.3
tqdm==4.48.2
urllib3==1.25.10
